Clean-ups and remove refresh one ix.
