+++ programs/scope/tests/tests_security_refresh_one.rs
@@ -7,36 +7,15 @@ use solana_program::{
     instruction::Instruction, sysvar::instructions::ID as SYSVAR_INSTRUCTIONS_ID,
 };
 use solana_program_test::tokio;
-use solana_sdk::{pubkey, signer::Signer};
+use solana_sdk::pubkey;
 use types::*;
 
-use crate::{
-    common::utils::AnchorErrorCode,
-    utils::{map_anchor_error, map_scope_error},
-};
+use crate::utils::map_scope_error;
 
-const TEST_PYTH_ORACLE: OracleConf = OracleConf {
-    pubkey: pubkey!("SomePythPriceAccount11111111111111111111111"),
-    token: 0,
-    price_type: TestOracleType::Pyth,
-    twap_enabled: false,
-    twap_source: None,
-};
-
-const TEST_PYTH2_ORACLE: OracleConf = OracleConf {
-    pubkey: pubkey!("SomePyth2PriceAccount1111111111111111111111"),
-    token: 1,
-    price_type: TestOracleType::Pyth,
-    twap_enabled: false,
-    twap_source: None,
-};
-
-// - [x] Wrong oracle mapping
-// - [x] Wrong oracle account (copy)
-// - [x] Wrong oracle account (mixing indexes)
-// - [x] Wrong sysvar instruction account
-// - [x] Instruction executed in CPI
-// - [x] Instruction preceded by non ComputeBudget instruction
+// Note: those tests aims to check the exact returned errors for a given price type while
+// `tests_security_refresh_list.rs` only check the absence of update because individual price
+// validation are not terminal unless only one price is being refreshed.
+// They are mostly kept as legacy from the old "RefreshOne" ix but can be interesting in some cases.
 
 // KTokens:
 // - [x] Wrong kToken additional global config account
@@ -52,312 +31,6 @@ const TEST_PYTH2_ORACLE: OracleConf = OracleConf {
 // - [x] Wrong Jupiter LP additional mint account
 // - [x] Wrong Jupiter LP additional custodies account
 // - [x] Wrong Jupiter LP additional oracles account
-
-#[tokio::test]
-async fn test_working_refresh_one() {
-    let (mut ctx, feed) =
-        fixtures::setup_scope(DEFAULT_FEED_NAME, vec![TEST_PYTH_ORACLE, TEST_PYTH2_ORACLE]).await;
-
-    // Change price
-    mock_oracles::set_price(
-        &mut ctx,
-        &feed,
-        &TEST_PYTH_ORACLE,
-        &Price { value: 1, exp: 6 },
-    )
-    .await;
-
-    // Refresh
-    let accounts = scope::accounts::RefreshOne {
-        oracle_prices: feed.prices,
-        oracle_mappings: feed.mapping,
-        oracle_twaps: feed.twaps,
-        instruction_sysvar_account_info: SYSVAR_INSTRUCTIONS_ID,
-        price_info: TEST_PYTH_ORACLE.pubkey,
-    };
-
-    let args = scope::instruction::RefreshOnePrice {
-        token: TEST_PYTH_ORACLE.token.try_into().unwrap(),
-    };
-
-    let ix = Instruction {
-        program_id: scope::id(),
-        accounts: accounts.to_account_metas(None),
-        data: args.data(),
-    };
-
-    ctx.send_transaction_with_bot(&[ix]).await.unwrap();
-
-    // Check price
-    let data: OraclePrices = ctx.get_zero_copy_account(&feed.prices).await.unwrap();
-    assert_eq!(data.prices[TEST_PYTH_ORACLE.token].price.value, 1);
-    assert_eq!(data.prices[TEST_PYTH_ORACLE.token].price.exp, 6);
-}
-
-// - [ ] Wrong oracle mapping
-#[tokio::test]
-async fn test_wrong_oracle_mapping() {
-    let (mut ctx, feed) = fixtures::setup_scope(DEFAULT_FEED_NAME, vec![TEST_PYTH_ORACLE]).await;
-
-    // Change price
-    mock_oracles::set_price(
-        &mut ctx,
-        &feed,
-        &TEST_PYTH_ORACLE,
-        &Price { value: 1, exp: 6 },
-    )
-    .await;
-
-    // Create a fake mapping account
-    let fake_mapping_pk = Pubkey::new_unique();
-    ctx.clone_account(&feed.mapping, &fake_mapping_pk).await;
-
-    // Refresh
-    let accounts = scope::accounts::RefreshOne {
-        oracle_prices: feed.prices,
-        oracle_mappings: fake_mapping_pk,
-        oracle_twaps: feed.twaps,
-        instruction_sysvar_account_info: SYSVAR_INSTRUCTIONS_ID,
-        price_info: TEST_PYTH_ORACLE.pubkey,
-    };
-
-    let args = scope::instruction::RefreshOnePrice {
-        token: TEST_PYTH_ORACLE.token.try_into().unwrap(),
-    };
-
-    let ix = Instruction {
-        program_id: scope::id(),
-        accounts: accounts.to_account_metas(None),
-        data: args.data(),
-    };
-
-    assert_eq!(
-        map_anchor_error(ctx.send_transaction_with_bot(&[ix]).await),
-        AnchorErrorCode::ConstraintHasOne,
-    );
-}
-
-// - [ ] Wrong oracle account (copy)
-#[tokio::test]
-async fn test_wrong_oracle_account_with_copy() {
-    let (mut ctx, feed) = fixtures::setup_scope(DEFAULT_FEED_NAME, vec![TEST_PYTH_ORACLE]).await;
-
-    // Change price
-    mock_oracles::set_price(
-        &mut ctx,
-        &feed,
-        &TEST_PYTH_ORACLE,
-        &Price { value: 1, exp: 6 },
-    )
-    .await;
-
-    // Create a fake mapping account
-    let fake_price_account = Pubkey::new_unique();
-    ctx.clone_account(&TEST_PYTH_ORACLE.pubkey, &fake_price_account)
-        .await;
-
-    // Refresh
-    let accounts = scope::accounts::RefreshOne {
-        oracle_prices: feed.prices,
-        oracle_mappings: feed.mapping,
-        oracle_twaps: feed.twaps,
-        instruction_sysvar_account_info: SYSVAR_INSTRUCTIONS_ID,
-        price_info: fake_price_account,
-    };
-
-    let args = scope::instruction::RefreshOnePrice {
-        token: TEST_PYTH_ORACLE.token.try_into().unwrap(),
-    };
-
-    let ix = Instruction {
-        program_id: scope::id(),
-        accounts: accounts.to_account_metas(None),
-        data: args.data(),
-    };
-
-    assert_eq!(
-        map_scope_error(ctx.send_transaction_with_bot(&[ix]).await),
-        ScopeError::UnexpectedAccount,
-    );
-}
-
-// - [ ] Wrong oracle account (mixing indexes)
-#[tokio::test]
-async fn test_wrong_index_oracle_account() {
-    let (mut ctx, feed) =
-        fixtures::setup_scope(DEFAULT_FEED_NAME, vec![TEST_PYTH_ORACLE, TEST_PYTH2_ORACLE]).await;
-
-    // Change price
-    mock_oracles::set_price(
-        &mut ctx,
-        &feed,
-        &TEST_PYTH_ORACLE,
-        &Price { value: 1, exp: 6 },
-    )
-    .await;
-
-    // Refresh
-    let accounts = scope::accounts::RefreshOne {
-        oracle_prices: feed.prices,
-        oracle_mappings: feed.mapping,
-        oracle_twaps: feed.twaps,
-        instruction_sysvar_account_info: SYSVAR_INSTRUCTIONS_ID,
-        price_info: TEST_PYTH_ORACLE.pubkey,
-    };
-
-    let args = scope::instruction::RefreshOnePrice { token: 1 };
-
-    let ix = Instruction {
-        program_id: scope::id(),
-        accounts: accounts.to_account_metas(None),
-        data: args.data(),
-    };
-
-    assert_eq!(
-        map_scope_error(ctx.send_transaction_with_bot(&[ix]).await),
-        ScopeError::UnexpectedAccount,
-    );
-}
-
-// - [ ] Wrong sysvar instruction account
-#[tokio::test]
-async fn test_wrong_sysvar_instructions() {
-    let (mut ctx, feed) =
-        fixtures::setup_scope(DEFAULT_FEED_NAME, vec![TEST_PYTH_ORACLE, TEST_PYTH2_ORACLE]).await;
-
-    // Change price
-    mock_oracles::set_price(
-        &mut ctx,
-        &feed,
-        &TEST_PYTH_ORACLE,
-        &Price { value: 1, exp: 6 },
-    )
-    .await;
-
-    // Create the fake sysvar
-    let wrong_sysvar_account = Pubkey::new_unique();
-
-    ctx.set_account(&wrong_sysvar_account, vec![0; 100], &Pubkey::new_unique());
-
-    // Refresh
-    let accounts = scope::accounts::RefreshOne {
-        oracle_prices: feed.prices,
-        oracle_mappings: feed.mapping,
-        oracle_twaps: feed.twaps,
-        instruction_sysvar_account_info: wrong_sysvar_account,
-        price_info: TEST_PYTH_ORACLE.pubkey,
-    };
-
-    let args = scope::instruction::RefreshOnePrice {
-        token: TEST_PYTH_ORACLE.token.try_into().unwrap(),
-    };
-
-    let ix = Instruction {
-        program_id: scope::id(),
-        accounts: accounts.to_account_metas(None),
-        data: args.data(),
-    };
-
-    let res = ctx.send_transaction_with_bot(&[ix]).await;
-    assert_eq!(map_anchor_error(res), AnchorErrorCode::ConstraintAddress);
-}
-
-// - [ ] Instruction executed in CPI
-#[tokio::test]
-async fn test_refresh_through_cpi() {
-    let (mut ctx, feed) =
-        fixtures::setup_scope(DEFAULT_FEED_NAME, vec![TEST_PYTH_ORACLE, TEST_PYTH2_ORACLE]).await;
-
-    // Change price
-    mock_oracles::set_price(
-        &mut ctx,
-        &feed,
-        &TEST_PYTH_ORACLE,
-        &Price { value: 1, exp: 6 },
-    )
-    .await;
-
-    // Refresh
-    let accounts = scope::accounts::RefreshOne {
-        oracle_prices: feed.prices,
-        oracle_mappings: feed.mapping,
-        oracle_twaps: feed.twaps,
-        instruction_sysvar_account_info: SYSVAR_INSTRUCTIONS_ID,
-        price_info: TEST_PYTH_ORACLE.pubkey,
-    };
-
-    let args = scope::instruction::RefreshOnePrice {
-        token: TEST_PYTH_ORACLE.token.try_into().unwrap(),
-    };
-
-    let ix = Instruction {
-        program_id: scope::id(),
-        accounts: accounts.to_account_metas(None),
-        data: args.data(),
-    };
-
-    let res = ctx.send_transaction_through_cpi(&[ix]).await;
-    assert_eq!(map_scope_error(res), ScopeError::RefreshInCPI);
-}
-
-// - [ ] Instruction preceded by non ComputeBudget instruction
-#[tokio::test]
-async fn test_refresh_with_unexpected_ix() {
-    let (mut ctx, feed) =
-        fixtures::setup_scope(DEFAULT_FEED_NAME, vec![TEST_PYTH_ORACLE, TEST_PYTH2_ORACLE]).await;
-
-    // Change price
-    mock_oracles::set_price(
-        &mut ctx,
-        &feed,
-        &TEST_PYTH_ORACLE,
-        &Price { value: 1, exp: 6 },
-    )
-    .await;
-
-    // Random update mapping as extra ix
-    let accounts = scope::accounts::UpdateOracleMapping {
-        admin: ctx.admin.pubkey(),
-        configuration: feed.conf,
-        oracle_mappings: feed.mapping,
-        price_info: Some(TEST_PYTH_ORACLE.pubkey),
-    };
-    let args = scope::instruction::UpdateMapping {
-        feed_name: feed.feed_name.clone(),
-        token: TEST_PYTH_ORACLE.token.try_into().unwrap(),
-        price_type: TEST_PYTH_ORACLE.price_type.to_u8(),
-        twap_enabled: false,
-        twap_source: u16::MAX,
-    };
-
-    let extra_ix = Instruction {
-        program_id: scope::id(),
-        accounts: accounts.to_account_metas(None),
-        data: args.data(),
-    };
-
-    // Refresh
-    let accounts = scope::accounts::RefreshOne {
-        oracle_prices: feed.prices,
-        oracle_mappings: feed.mapping,
-        oracle_twaps: feed.twaps,
-        instruction_sysvar_account_info: SYSVAR_INSTRUCTIONS_ID,
-        price_info: TEST_PYTH_ORACLE.pubkey,
-    };
-
-    let args = scope::instruction::RefreshOnePrice {
-        token: TEST_PYTH_ORACLE.token.try_into().unwrap(),
-    };
-
-    let ix = Instruction {
-        program_id: scope::id(),
-        accounts: accounts.to_account_metas(None),
-        data: args.data(),
-    };
-
-    let res = ctx.send_transaction(&[extra_ix, ix]).await;
-    assert_eq!(map_scope_error(res), ScopeError::RefreshWithUnexpectedIxs);
-}
 
 #[cfg(feature = "yvaults")]
 mod ktoken_tests {
@@ -398,20 +71,19 @@ mod ktoken_tests {
         .await;
 
         // Refresh
-        let mut accounts = scope::accounts::RefreshOne {
+        let mut accounts = scope::accounts::RefreshList {
             oracle_prices: feed.prices,
             oracle_mappings: feed.mapping,
             oracle_twaps: feed.twaps,
             instruction_sysvar_account_info: SYSVAR_INSTRUCTIONS_ID,
-            price_info: TEST_ORCA_KTOKEN_ORACLE.pubkey,
         }
         .to_account_metas(None);
         let mut refresh_accounts =
-            utils::get_remaining_accounts(&mut ctx, &TEST_ORCA_KTOKEN_ORACLE).await;
+            utils::get_refresh_list_accounts(&mut ctx, &TEST_ORCA_KTOKEN_ORACLE).await;
         accounts.append(&mut refresh_accounts);
 
-        let args = scope::instruction::RefreshOnePrice {
-            token: TEST_ORCA_KTOKEN_ORACLE.token.try_into().unwrap(),
+        let args = scope::instruction::RefreshPriceList {
+            tokens: vec![TEST_ORCA_KTOKEN_ORACLE.token.try_into().unwrap()],
         };
 
         let ix = Instruction {
@@ -446,20 +118,19 @@ mod ktoken_tests {
         .await;
 
         // Refresh
-        let mut accounts = scope::accounts::RefreshOne {
+        let mut accounts = scope::accounts::RefreshList {
             oracle_prices: feed.prices,
             oracle_mappings: feed.mapping,
             oracle_twaps: feed.twaps,
             instruction_sysvar_account_info: SYSVAR_INSTRUCTIONS_ID,
-            price_info: TEST_RAYDIUM_KTOKEN_ORACLE.pubkey,
         }
         .to_account_metas(None);
         let mut refresh_accounts =
-            utils::get_remaining_accounts(&mut ctx, &TEST_RAYDIUM_KTOKEN_ORACLE).await;
+            utils::get_refresh_list_accounts(&mut ctx, &TEST_RAYDIUM_KTOKEN_ORACLE).await;
         accounts.append(&mut refresh_accounts);
 
-        let args = scope::instruction::RefreshOnePrice {
-            token: TEST_RAYDIUM_KTOKEN_ORACLE.token.try_into().unwrap(),
+        let args = scope::instruction::RefreshPriceList {
+            tokens: vec![TEST_RAYDIUM_KTOKEN_ORACLE.token.try_into().unwrap()],
         };
 
         let ix = Instruction {
@@ -505,16 +176,15 @@ mod ktoken_tests {
             .await;
 
         // Refresh
-        let mut accounts = scope::accounts::RefreshOne {
+        let mut accounts = scope::accounts::RefreshList {
             oracle_prices: feed.prices,
             oracle_mappings: feed.mapping,
             oracle_twaps: feed.twaps,
             instruction_sysvar_account_info: SYSVAR_INSTRUCTIONS_ID,
-            price_info: TEST_ORCA_KTOKEN_ORACLE.pubkey,
         }
         .to_account_metas(None);
         let mut refresh_accounts =
-            utils::get_remaining_accounts(&mut ctx, &TEST_ORCA_KTOKEN_ORACLE).await;
+            utils::get_refresh_list_accounts(&mut ctx, &TEST_ORCA_KTOKEN_ORACLE).await;
         accounts.append(&mut refresh_accounts);
         // Set the wrong global config
         accounts.iter_mut().for_each(|account| {
@@ -523,8 +193,8 @@ mod ktoken_tests {
             }
         });
 
-        let args = scope::instruction::RefreshOnePrice {
-            token: TEST_ORCA_KTOKEN_ORACLE.token.try_into().unwrap(),
+        let args = scope::instruction::RefreshPriceList {
+            tokens: vec![TEST_ORCA_KTOKEN_ORACLE.token.try_into().unwrap()],
         };
 
         let ix = Instruction {
@@ -563,16 +233,15 @@ mod ktoken_tests {
             .await;
 
         // Refresh
-        let mut accounts = scope::accounts::RefreshOne {
+        let mut accounts = scope::accounts::RefreshList {
             oracle_prices: feed.prices,
             oracle_mappings: feed.mapping,
             oracle_twaps: feed.twaps,
             instruction_sysvar_account_info: SYSVAR_INSTRUCTIONS_ID,
-            price_info: TEST_RAYDIUM_KTOKEN_ORACLE.pubkey,
         }
         .to_account_metas(None);
         let mut refresh_accounts =
-            utils::get_remaining_accounts(&mut ctx, &TEST_RAYDIUM_KTOKEN_ORACLE).await;
+            utils::get_refresh_list_accounts(&mut ctx, &TEST_RAYDIUM_KTOKEN_ORACLE).await;
         accounts.append(&mut refresh_accounts);
         // Set the wrong global config
         accounts.iter_mut().for_each(|account| {
@@ -581,8 +250,8 @@ mod ktoken_tests {
             }
         });
 
-        let args = scope::instruction::RefreshOnePrice {
-            token: TEST_RAYDIUM_KTOKEN_ORACLE.token.try_into().unwrap(),
+        let args = scope::instruction::RefreshPriceList {
+            tokens: vec![TEST_RAYDIUM_KTOKEN_ORACLE.token.try_into().unwrap()],
         };
 
         let ix = Instruction {
@@ -625,16 +294,15 @@ mod ktoken_tests {
             .await;
 
         // Refresh
-        let mut accounts = scope::accounts::RefreshOne {
+        let mut accounts = scope::accounts::RefreshList {
             oracle_prices: feed.prices,
             oracle_mappings: feed.mapping,
             oracle_twaps: feed.twaps,
             instruction_sysvar_account_info: SYSVAR_INSTRUCTIONS_ID,
-            price_info: TEST_ORCA_KTOKEN_ORACLE.pubkey,
         }
         .to_account_metas(None);
         let mut refresh_accounts =
-            utils::get_remaining_accounts(&mut ctx, &TEST_ORCA_KTOKEN_ORACLE).await;
+            utils::get_refresh_list_accounts(&mut ctx, &TEST_ORCA_KTOKEN_ORACLE).await;
         accounts.append(&mut refresh_accounts);
         // Set the wrong collateral infos
         accounts.iter_mut().for_each(|account| {
@@ -643,8 +311,8 @@ mod ktoken_tests {
             }
         });
 
-        let args = scope::instruction::RefreshOnePrice {
-            token: TEST_ORCA_KTOKEN_ORACLE.token.try_into().unwrap(),
+        let args = scope::instruction::RefreshPriceList {
+            tokens: vec![TEST_ORCA_KTOKEN_ORACLE.token.try_into().unwrap()],
         };
 
         let ix = Instruction {
@@ -687,16 +355,15 @@ mod ktoken_tests {
             .await;
 
         // Refresh
-        let mut accounts = scope::accounts::RefreshOne {
+        let mut accounts = scope::accounts::RefreshList {
             oracle_prices: feed.prices,
             oracle_mappings: feed.mapping,
             oracle_twaps: feed.twaps,
             instruction_sysvar_account_info: SYSVAR_INSTRUCTIONS_ID,
-            price_info: TEST_RAYDIUM_KTOKEN_ORACLE.pubkey,
         }
         .to_account_metas(None);
         let mut refresh_accounts =
-            utils::get_remaining_accounts(&mut ctx, &TEST_RAYDIUM_KTOKEN_ORACLE).await;
+            utils::get_refresh_list_accounts(&mut ctx, &TEST_RAYDIUM_KTOKEN_ORACLE).await;
         accounts.append(&mut refresh_accounts);
         // Set the wrong collateral infos
         accounts.iter_mut().for_each(|account| {
@@ -705,8 +372,8 @@ mod ktoken_tests {
             }
         });
 
-        let args = scope::instruction::RefreshOnePrice {
-            token: TEST_RAYDIUM_KTOKEN_ORACLE.token.try_into().unwrap(),
+        let args = scope::instruction::RefreshPriceList {
+            tokens: vec![TEST_RAYDIUM_KTOKEN_ORACLE.token.try_into().unwrap()],
         };
 
         let ix = Instruction {
@@ -745,16 +412,15 @@ mod ktoken_tests {
             .await;
 
         // Refresh
-        let mut accounts = scope::accounts::RefreshOne {
+        let mut accounts = scope::accounts::RefreshList {
             oracle_prices: feed.prices,
             oracle_mappings: feed.mapping,
             oracle_twaps: feed.twaps,
             instruction_sysvar_account_info: SYSVAR_INSTRUCTIONS_ID,
-            price_info: TEST_ORCA_KTOKEN_ORACLE.pubkey,
         }
         .to_account_metas(None);
         let mut refresh_accounts =
-            utils::get_remaining_accounts(&mut ctx, &TEST_ORCA_KTOKEN_ORACLE).await;
+            utils::get_refresh_list_accounts(&mut ctx, &TEST_ORCA_KTOKEN_ORACLE).await;
         accounts.append(&mut refresh_accounts);
         // Set the wrong orca whirlpool
         accounts.iter_mut().for_each(|account| {
@@ -763,8 +429,8 @@ mod ktoken_tests {
             }
         });
 
-        let args = scope::instruction::RefreshOnePrice {
-            token: TEST_ORCA_KTOKEN_ORACLE.token.try_into().unwrap(),
+        let args = scope::instruction::RefreshPriceList {
+            tokens: vec![TEST_ORCA_KTOKEN_ORACLE.token.try_into().unwrap()],
         };
 
         let ix = Instruction {
@@ -802,16 +468,15 @@ mod ktoken_tests {
         ctx.clone_account(&strategy.pool, &wrong_raydium_pool).await;
 
         // Refresh
-        let mut accounts = scope::accounts::RefreshOne {
+        let mut accounts = scope::accounts::RefreshList {
             oracle_prices: feed.prices,
             oracle_mappings: feed.mapping,
             oracle_twaps: feed.twaps,
             instruction_sysvar_account_info: SYSVAR_INSTRUCTIONS_ID,
-            price_info: TEST_RAYDIUM_KTOKEN_ORACLE.pubkey,
         }
         .to_account_metas(None);
         let mut refresh_accounts =
-            utils::get_remaining_accounts(&mut ctx, &TEST_RAYDIUM_KTOKEN_ORACLE).await;
+            utils::get_refresh_list_accounts(&mut ctx, &TEST_RAYDIUM_KTOKEN_ORACLE).await;
         accounts.append(&mut refresh_accounts);
         // Set the wrong orca whirlpool
         accounts.iter_mut().for_each(|account| {
@@ -820,8 +485,8 @@ mod ktoken_tests {
             }
         });
 
-        let args = scope::instruction::RefreshOnePrice {
-            token: TEST_RAYDIUM_KTOKEN_ORACLE.token.try_into().unwrap(),
+        let args = scope::instruction::RefreshPriceList {
+            tokens: vec![TEST_RAYDIUM_KTOKEN_ORACLE.token.try_into().unwrap()],
         };
 
         let ix = Instruction {
@@ -860,16 +525,15 @@ mod ktoken_tests {
             .await;
 
         // Refresh
-        let mut accounts = scope::accounts::RefreshOne {
+        let mut accounts = scope::accounts::RefreshList {
             oracle_prices: feed.prices,
             oracle_mappings: feed.mapping,
             oracle_twaps: feed.twaps,
             instruction_sysvar_account_info: SYSVAR_INSTRUCTIONS_ID,
-            price_info: TEST_ORCA_KTOKEN_ORACLE.pubkey,
         }
         .to_account_metas(None);
         let mut refresh_accounts =
-            utils::get_remaining_accounts(&mut ctx, &TEST_ORCA_KTOKEN_ORACLE).await;
+            utils::get_refresh_list_accounts(&mut ctx, &TEST_ORCA_KTOKEN_ORACLE).await;
         accounts.append(&mut refresh_accounts);
         // Set the wrong orca position
         accounts.iter_mut().for_each(|account| {
@@ -878,8 +542,8 @@ mod ktoken_tests {
             }
         });
 
-        let args = scope::instruction::RefreshOnePrice {
-            token: TEST_ORCA_KTOKEN_ORACLE.token.try_into().unwrap(),
+        let args = scope::instruction::RefreshPriceList {
+            tokens: vec![TEST_ORCA_KTOKEN_ORACLE.token.try_into().unwrap()],
         };
 
         let ix = Instruction {
@@ -918,16 +582,15 @@ mod ktoken_tests {
             .await;
 
         // Refresh
-        let mut accounts = scope::accounts::RefreshOne {
+        let mut accounts = scope::accounts::RefreshList {
             oracle_prices: feed.prices,
             oracle_mappings: feed.mapping,
             oracle_twaps: feed.twaps,
             instruction_sysvar_account_info: SYSVAR_INSTRUCTIONS_ID,
-            price_info: TEST_RAYDIUM_KTOKEN_ORACLE.pubkey,
         }
         .to_account_metas(None);
         let mut refresh_accounts =
-            utils::get_remaining_accounts(&mut ctx, &TEST_RAYDIUM_KTOKEN_ORACLE).await;
+            utils::get_refresh_list_accounts(&mut ctx, &TEST_RAYDIUM_KTOKEN_ORACLE).await;
         accounts.append(&mut refresh_accounts);
         // Set the wrong orca position
         accounts.iter_mut().for_each(|account| {
@@ -936,8 +599,8 @@ mod ktoken_tests {
             }
         });
 
-        let args = scope::instruction::RefreshOnePrice {
-            token: TEST_RAYDIUM_KTOKEN_ORACLE.token.try_into().unwrap(),
+        let args = scope::instruction::RefreshPriceList {
+            tokens: vec![TEST_RAYDIUM_KTOKEN_ORACLE.token.try_into().unwrap()],
         };
 
         let ix = Instruction {
@@ -976,16 +639,15 @@ mod ktoken_tests {
             .await;
 
         // Refresh
-        let mut accounts = scope::accounts::RefreshOne {
+        let mut accounts = scope::accounts::RefreshList {
             oracle_prices: feed.prices,
             oracle_mappings: feed.mapping,
             oracle_twaps: feed.twaps,
             instruction_sysvar_account_info: SYSVAR_INSTRUCTIONS_ID,
-            price_info: TEST_ORCA_KTOKEN_ORACLE.pubkey,
         }
         .to_account_metas(None);
         let mut refresh_accounts =
-            utils::get_remaining_accounts(&mut ctx, &TEST_ORCA_KTOKEN_ORACLE).await;
+            utils::get_refresh_list_accounts(&mut ctx, &TEST_ORCA_KTOKEN_ORACLE).await;
         // Set the wrong scope prices
         refresh_accounts.iter_mut().for_each(|account| {
             if account.pubkey == strategy.scope_prices {
@@ -994,8 +656,8 @@ mod ktoken_tests {
         });
         accounts.append(&mut refresh_accounts);
 
-        let args = scope::instruction::RefreshOnePrice {
-            token: TEST_ORCA_KTOKEN_ORACLE.token.try_into().unwrap(),
+        let args = scope::instruction::RefreshPriceList {
+            tokens: vec![TEST_ORCA_KTOKEN_ORACLE.token.try_into().unwrap()],
         };
 
         let ix = Instruction {
@@ -1034,16 +696,15 @@ mod ktoken_tests {
             .await;
 
         // Refresh
-        let mut accounts = scope::accounts::RefreshOne {
+        let mut accounts = scope::accounts::RefreshList {
             oracle_prices: feed.prices,
             oracle_mappings: feed.mapping,
             oracle_twaps: feed.twaps,
             instruction_sysvar_account_info: SYSVAR_INSTRUCTIONS_ID,
-            price_info: TEST_RAYDIUM_KTOKEN_ORACLE.pubkey,
         }
         .to_account_metas(None);
         let mut refresh_accounts =
-            utils::get_remaining_accounts(&mut ctx, &TEST_RAYDIUM_KTOKEN_ORACLE).await;
+            utils::get_refresh_list_accounts(&mut ctx, &TEST_RAYDIUM_KTOKEN_ORACLE).await;
         refresh_accounts.iter_mut().for_each(|account| {
             if account.pubkey == strategy.scope_prices {
                 account.pubkey = wrong_scope_prices;
@@ -1051,8 +712,8 @@ mod ktoken_tests {
         });
         accounts.append(&mut refresh_accounts);
 
-        let args = scope::instruction::RefreshOnePrice {
-            token: TEST_RAYDIUM_KTOKEN_ORACLE.token.try_into().unwrap(),
+        let args = scope::instruction::RefreshPriceList {
+            tokens: vec![TEST_RAYDIUM_KTOKEN_ORACLE.token.try_into().unwrap()],
         };
 
         let ix = Instruction {
@@ -1089,19 +750,18 @@ mod test_jlp_fetch {
         mock_oracles::set_price(&mut ctx, &feed, &TEST_JLP_ORACLE, &price).await;
 
         // Refresh
-        let mut accounts = scope::accounts::RefreshOne {
+        let mut accounts = scope::accounts::RefreshList {
             oracle_prices: feed.prices,
             oracle_mappings: feed.mapping,
             oracle_twaps: feed.twaps,
             instruction_sysvar_account_info: SYSVAR_INSTRUCTIONS_ID,
-            price_info: TEST_JLP_ORACLE.pubkey,
         }
         .to_account_metas(None);
 
-        accounts.append(&mut utils::get_remaining_accounts(&mut ctx, &TEST_JLP_ORACLE).await);
+        accounts.append(&mut utils::get_refresh_list_accounts(&mut ctx, &TEST_JLP_ORACLE).await);
 
-        let args = scope::instruction::RefreshOnePrice {
-            token: TEST_JLP_ORACLE.token.try_into().unwrap(),
+        let args = scope::instruction::RefreshPriceList {
+            tokens: vec![TEST_JLP_ORACLE.token.try_into().unwrap()],
         };
 
         let ix = Instruction {
@@ -1129,17 +789,16 @@ mod test_jlp_fetch {
         mock_oracles::set_price(&mut ctx, &feed, &TEST_JLP_ORACLE, &price).await;
 
         // Refresh
-        let mut accounts = scope::accounts::RefreshOne {
+        let mut accounts = scope::accounts::RefreshList {
             oracle_prices: feed.prices,
             oracle_mappings: feed.mapping,
             oracle_twaps: feed.twaps,
             instruction_sysvar_account_info: SYSVAR_INSTRUCTIONS_ID,
-            price_info: TEST_JLP_ORACLE.pubkey,
         }
         .to_account_metas(None);
 
         let mut remaining_accounts =
-            utils::get_remaining_accounts(&mut ctx, &TEST_JLP_ORACLE).await;
+            utils::get_refresh_list_accounts(&mut ctx, &TEST_JLP_ORACLE).await;
 
         let mint = &mut remaining_accounts[0];
         let mint_pk = mint.pubkey;
@@ -1148,8 +807,8 @@ mod test_jlp_fetch {
         mint.pubkey = cloned_mint;
         accounts.append(&mut remaining_accounts);
 
-        let args = scope::instruction::RefreshOnePrice {
-            token: TEST_JLP_ORACLE.token.try_into().unwrap(),
+        let args = scope::instruction::RefreshPriceList {
+            tokens: vec![TEST_JLP_ORACLE.token.try_into().unwrap()],
         };
 
         let ix = Instruction {
@@ -1187,19 +846,18 @@ mod test_jlp_compute {
         mock_oracles::set_price(&mut ctx, &feed, &TEST_JLP_ORACLE, &price).await;
 
         // Refresh
-        let mut accounts = scope::accounts::RefreshOne {
+        let mut accounts = scope::accounts::RefreshList {
             oracle_prices: feed.prices,
             oracle_mappings: feed.mapping,
             oracle_twaps: feed.twaps,
             instruction_sysvar_account_info: SYSVAR_INSTRUCTIONS_ID,
-            price_info: TEST_JLP_ORACLE.pubkey,
         }
         .to_account_metas(None);
 
-        accounts.append(&mut utils::get_remaining_accounts(&mut ctx, &TEST_JLP_ORACLE).await);
+        accounts.append(&mut utils::get_refresh_list_accounts(&mut ctx, &TEST_JLP_ORACLE).await);
 
-        let args = scope::instruction::RefreshOnePrice {
-            token: TEST_JLP_ORACLE.token.try_into().unwrap(),
+        let args = scope::instruction::RefreshPriceList {
+            tokens: vec![TEST_JLP_ORACLE.token.try_into().unwrap()],
         };
 
         let ix = Instruction {
@@ -1237,16 +895,15 @@ mod test_jlp_compute {
         mock_oracles::set_price(&mut ctx, &feed, &TEST_JLP_ORACLE, &price).await;
 
         // Refresh
-        let mut accounts = scope::accounts::RefreshOne {
+        let mut accounts = scope::accounts::RefreshList {
             oracle_prices: feed.prices,
             oracle_mappings: feed.mapping,
             oracle_twaps: feed.twaps,
             instruction_sysvar_account_info: SYSVAR_INSTRUCTIONS_ID,
-            price_info: TEST_JLP_ORACLE.pubkey,
         }
         .to_account_metas(None);
 
-        let remaining_accounts = utils::get_remaining_accounts(&mut ctx, &TEST_JLP_ORACLE).await;
+        let remaining_accounts = utils::get_refresh_list_accounts(&mut ctx, &TEST_JLP_ORACLE).await;
 
         for i in 0..remaining_accounts.len() {
             let mut remaining_accounts = remaining_accounts.clone();
@@ -1257,8 +914,8 @@ mod test_jlp_compute {
             acc.pubkey = cloned_acc;
             accounts.append(&mut remaining_accounts);
 
-            let args = scope::instruction::RefreshOnePrice {
-                token: TEST_JLP_ORACLE.token.try_into().unwrap(),
+            let args = scope::instruction::RefreshPriceList {
+                tokens: vec![TEST_JLP_ORACLE.token.try_into().unwrap()],
             };
 
             let ix = Instruction {
