+++ programs/scope/tests/tests_twap.rs
@@ -1,6 +1,6 @@
 mod common;
 
-use crate::common::client::refresh_one_ix;
+use crate::common::client::refresh_simple_oracle_ix;
 use crate::common::utils::map_scope_error;
 use crate::{client::reset_twap, common::fixtures::setup_mapping_for_token_with_twap};
 use anchor_lang::{InstructionData, ToAccountMetas};
@@ -131,7 +131,7 @@ async fn test_set_price_sets_initial_twap() {
     setup_mapping_for_token_with_twap(&mut ctx, &feed, TEST_PYTH_ORACLE, TEST_TWAP).await;
 
     // Refresh
-    let refresh_ix = refresh_one_ix(&feed, TEST_PYTH_ORACLE);
+    let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_PYTH_ORACLE);
     ctx.send_transaction_with_bot(&[refresh_ix]).await.unwrap();
 
     {
@@ -148,7 +148,7 @@ async fn test_set_price_sets_initial_twap() {
     }
 
     {
-        let refresh_ix = refresh_one_ix(&feed, TEST_TWAP);
+        let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_TWAP);
         let res = ctx.send_transaction_with_bot(&[refresh_ix]).await;
 
         assert_eq!(
@@ -169,7 +169,7 @@ async fn test_2_prices_with_same_value_no_twap_change() {
     setup_mapping_for_token_with_twap(&mut ctx, &feed, TEST_PYTH_ORACLE, TEST_TWAP).await;
 
     // Refresh price
-    let refresh_ix = refresh_one_ix(&feed, TEST_PYTH_ORACLE);
+    let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_PYTH_ORACLE);
     ctx.send_transaction_with_bot(&[refresh_ix]).await.unwrap();
 
     // Verify that TWAP is the same as the price as it is the first sample
@@ -188,7 +188,7 @@ async fn test_2_prices_with_same_value_no_twap_change() {
     // Fast forward not enough time and refresh price with the same value
     ctx.fast_forward_seconds(10).await;
     mock_oracles::set_price(&mut ctx, &feed, &TEST_PYTH_ORACLE, &token_price).await;
-    let refresh_ix = refresh_one_ix(&feed, TEST_PYTH_ORACLE);
+    let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_PYTH_ORACLE);
     ctx.send_transaction_with_bot(&[refresh_ix]).await.unwrap();
 
     // verify that the twap value and ts didn't change (too frequent updates)
@@ -211,7 +211,7 @@ async fn test_2_prices_with_same_value_no_twap_change() {
     // Fast forward time and refresh price with the same value
     ctx.fast_forward_seconds(90).await;
     mock_oracles::set_price(&mut ctx, &feed, &TEST_PYTH_ORACLE, &token_price).await;
-    let refresh_ix = refresh_one_ix(&feed, TEST_PYTH_ORACLE);
+    let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_PYTH_ORACLE);
     ctx.send_transaction_with_bot(&[refresh_ix]).await.unwrap();
 
     // verify that the twap value didn't change but the twap date increased (update after right amount of time)
@@ -243,7 +243,7 @@ async fn test_2_prices_with_same_value_no_twap_change_1h() {
     setup_mapping_for_token_with_twap(&mut ctx, &feed, TEST_PYTH_ORACLE, TEST_TWAP).await;
 
     // Refresh price
-    let refresh_ix = refresh_one_ix(&feed, TEST_PYTH_ORACLE);
+    let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_PYTH_ORACLE);
     ctx.send_transaction_with_bot(&[refresh_ix]).await.unwrap();
 
     // Verify that TWAP is the same as the price as it is the first sample
@@ -262,7 +262,7 @@ async fn test_2_prices_with_same_value_no_twap_change_1h() {
     // Fast forward time and refresh price with the same value
     ctx.fast_forward_seconds(60 * 60 + 120).await; // 1h
     mock_oracles::set_price(&mut ctx, &feed, &TEST_PYTH_ORACLE, &token_price).await;
-    let refresh_ix = refresh_one_ix(&feed, TEST_PYTH_ORACLE);
+    let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_PYTH_ORACLE);
     ctx.send_transaction_with_bot(&[refresh_ix]).await.unwrap();
 
     // verify that the twap value didn't change but the twap date increased
@@ -284,7 +284,7 @@ async fn test_2_prices_with_same_value_no_twap_change_1h() {
 
     // verify that this is not enough samples to use this twap
     {
-        let refresh_ix = refresh_one_ix(&feed, TEST_TWAP);
+        let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_TWAP);
         let res = ctx.send_transaction_with_bot(&[refresh_ix]).await;
 
         assert_eq!(
@@ -305,7 +305,7 @@ async fn test_9_is_not_enough_twap_samples() {
     setup_mapping_for_token_with_twap(&mut ctx, &feed, TEST_PYTH_ORACLE, TEST_TWAP).await;
 
     // Refresh price
-    let refresh_ix = refresh_one_ix(&feed, TEST_PYTH_ORACLE);
+    let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_PYTH_ORACLE);
     ctx.send_transaction_with_bot(&[refresh_ix]).await.unwrap();
 
     // Verify that TWAP is the same as the price as it is the first sample
@@ -325,7 +325,7 @@ async fn test_9_is_not_enough_twap_samples() {
     for _ in 0..9 {
         ctx.fast_forward_seconds(60 * 60 / 9).await;
         mock_oracles::set_price(&mut ctx, &feed, &TEST_PYTH_ORACLE, &token_price).await;
-        let refresh_ix = refresh_one_ix(&feed, TEST_PYTH_ORACLE);
+        let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_PYTH_ORACLE);
         ctx.send_transaction_with_bot(&[refresh_ix]).await.unwrap();
     }
 
@@ -348,7 +348,7 @@ async fn test_9_is_not_enough_twap_samples() {
 
     // verify that this is not enough samples to use this twap
     {
-        let refresh_ix = refresh_one_ix(&feed, TEST_TWAP);
+        let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_TWAP);
         let res = ctx.send_transaction_with_bot(&[refresh_ix]).await;
 
         assert_eq!(
@@ -369,7 +369,7 @@ async fn test_not_enough_twap_samples_at_period_start() {
     setup_mapping_for_token_with_twap(&mut ctx, &feed, TEST_PYTH_ORACLE, TEST_TWAP).await;
 
     // Refresh price
-    let refresh_ix = refresh_one_ix(&feed, TEST_PYTH_ORACLE);
+    let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_PYTH_ORACLE);
     ctx.send_transaction_with_bot(&[refresh_ix]).await.unwrap();
 
     // Verify that TWAP is the same as the price as it is the first sample
@@ -392,7 +392,7 @@ async fn test_not_enough_twap_samples_at_period_start() {
     for _ in 0..20 {
         ctx.fast_forward_seconds(60 * 40 / 20).await;
         mock_oracles::set_price(&mut ctx, &feed, &TEST_PYTH_ORACLE, &token_price).await;
-        let refresh_ix = refresh_one_ix(&feed, TEST_PYTH_ORACLE);
+        let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_PYTH_ORACLE);
         ctx.send_transaction_with_bot(&[refresh_ix]).await.unwrap();
     }
 
@@ -415,7 +415,7 @@ async fn test_not_enough_twap_samples_at_period_start() {
 
     // verify that this is not enough samples to use this twap
     {
-        let refresh_ix = refresh_one_ix(&feed, TEST_TWAP);
+        let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_TWAP);
         let res = ctx.send_transaction_with_bot(&[refresh_ix]).await;
 
         assert_eq!(
@@ -436,7 +436,7 @@ async fn test_not_enough_twap_samples_at_period_end() {
     setup_mapping_for_token_with_twap(&mut ctx, &feed, TEST_PYTH_ORACLE, TEST_TWAP).await;
 
     // Refresh price
-    let refresh_ix = refresh_one_ix(&feed, TEST_PYTH_ORACLE);
+    let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_PYTH_ORACLE);
     ctx.send_transaction_with_bot(&[refresh_ix]).await.unwrap();
 
     // Verify that TWAP is the same as the price as it is the first sample
@@ -456,7 +456,7 @@ async fn test_not_enough_twap_samples_at_period_end() {
     for _ in 0..20 {
         ctx.fast_forward_seconds(60 * 39 / 20).await;
         mock_oracles::set_price(&mut ctx, &feed, &TEST_PYTH_ORACLE, &token_price).await;
-        let refresh_ix = refresh_one_ix(&feed, TEST_PYTH_ORACLE);
+        let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_PYTH_ORACLE);
         ctx.send_transaction_with_bot(&[refresh_ix]).await.unwrap();
     }
 
@@ -482,7 +482,7 @@ async fn test_not_enough_twap_samples_at_period_end() {
 
     // verify that this is not enough samples to use this twap
     {
-        let refresh_ix = refresh_one_ix(&feed, TEST_TWAP);
+        let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_TWAP);
         let res = ctx.send_transaction_with_bot(&[refresh_ix]).await;
 
         assert_eq!(
@@ -506,7 +506,7 @@ async fn test_multiple_prices_with_same_value_no_twap_change() {
     setup_mapping_for_token_with_twap(&mut ctx, &feed, TEST_PYTH_ORACLE, TEST_TWAP).await;
 
     // Refresh price
-    let refresh_ix = refresh_one_ix(&feed, TEST_PYTH_ORACLE);
+    let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_PYTH_ORACLE);
     ctx.send_transaction_with_bot(&[refresh_ix]).await.unwrap();
 
     // Verify that TWAP is the same as the price as it is the first sample
@@ -523,7 +523,7 @@ async fn test_multiple_prices_with_same_value_no_twap_change() {
         // Fast forward time and refresh price with the same value
         ctx.fast_forward_seconds(60).await;
         mock_oracles::set_price(&mut ctx, &feed, &TEST_PYTH_ORACLE, &token_price).await;
-        let refresh_ix = refresh_one_ix(&feed, TEST_PYTH_ORACLE);
+        let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_PYTH_ORACLE);
         ctx.send_transaction_with_bot(&[refresh_ix]).await.unwrap();
 
         // verify that the twap value didn't change but the twap date increased
@@ -543,7 +543,7 @@ async fn test_multiple_prices_with_same_value_no_twap_change() {
     }
 
     {
-        let refresh_ix = refresh_one_ix(&feed, TEST_TWAP);
+        let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_TWAP);
         ctx.send_transaction_with_bot(&[refresh_ix]).await.unwrap();
 
         let oracle_prices: OraclePrices = ctx.get_zero_copy_account(&feed.prices).await.unwrap();
@@ -570,7 +570,7 @@ async fn test_multiple_prices_with_same_increasing_value_twap_increases() {
     setup_mapping_for_token_with_twap(&mut ctx, &feed, TEST_PYTH_ORACLE, TEST_TWAP).await;
 
     // Refresh price
-    let refresh_ix = refresh_one_ix(&feed, TEST_PYTH_ORACLE);
+    let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_PYTH_ORACLE);
     ctx.send_transaction_with_bot(&[refresh_ix]).await.unwrap();
 
     // Verify that TWAP is the same as the price as it is the first sample
@@ -589,7 +589,7 @@ async fn test_multiple_prices_with_same_increasing_value_twap_increases() {
         // Fast forward time and refresh price with the new value
         ctx.fast_forward_seconds(60).await;
         mock_oracles::set_price(&mut ctx, &feed, &TEST_PYTH_ORACLE, &token_price).await;
-        let refresh_ix = refresh_one_ix(&feed, TEST_PYTH_ORACLE);
+        let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_PYTH_ORACLE);
         ctx.send_transaction_with_bot(&[refresh_ix]).await.unwrap();
 
         // verify that the twap value didn't change but the twap date increased
@@ -617,12 +617,12 @@ async fn test_multiple_prices_with_same_increasing_value_twap_increases() {
     for _ in 1..60 {
         ctx.fast_forward_seconds(60).await;
         mock_oracles::set_price(&mut ctx, &feed, &TEST_PYTH_ORACLE, &token_price).await;
-        let refresh_ix = refresh_one_ix(&feed, TEST_PYTH_ORACLE);
+        let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_PYTH_ORACLE);
         ctx.send_transaction_with_bot(&[refresh_ix]).await.unwrap();
     }
 
     {
-        let refresh_ix = refresh_one_ix(&feed, TEST_TWAP);
+        let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_TWAP);
         ctx.send_transaction_with_bot(&[refresh_ix]).await.unwrap();
 
         let oracle_prices: OraclePrices = ctx.get_zero_copy_account(&feed.prices).await.unwrap();
@@ -654,7 +654,7 @@ async fn test_multiple_prices_with_decreasing_value_twap_decreases() {
     setup_mapping_for_token_with_twap(&mut ctx, &feed, TEST_PYTH_ORACLE, TEST_TWAP).await;
 
     // Refresh price
-    let refresh_ix = refresh_one_ix(&feed, TEST_PYTH_ORACLE);
+    let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_PYTH_ORACLE);
     ctx.send_transaction_with_bot(&[refresh_ix]).await.unwrap();
 
     // Verify that TWAP is the same as the price as it is the first sample
@@ -673,7 +673,7 @@ async fn test_multiple_prices_with_decreasing_value_twap_decreases() {
         // Fast forward time and refresh price with the new value
         ctx.fast_forward_seconds(70).await;
         mock_oracles::set_price(&mut ctx, &feed, &TEST_PYTH_ORACLE, &token_price).await;
-        let refresh_ix = refresh_one_ix(&feed, TEST_PYTH_ORACLE);
+        let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_PYTH_ORACLE);
         ctx.send_transaction_with_bot(&[refresh_ix]).await.unwrap();
 
         let oracle_twaps_updated: Box<OracleTwaps> =
@@ -700,12 +700,12 @@ async fn test_multiple_prices_with_decreasing_value_twap_decreases() {
     for _ in 1..60 {
         ctx.fast_forward_seconds(60).await;
         mock_oracles::set_price(&mut ctx, &feed, &TEST_PYTH_ORACLE, &token_price).await;
-        let refresh_ix = refresh_one_ix(&feed, TEST_PYTH_ORACLE);
+        let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_PYTH_ORACLE);
         ctx.send_transaction_with_bot(&[refresh_ix]).await.unwrap();
     }
 
     {
-        let refresh_ix = refresh_one_ix(&feed, TEST_TWAP);
+        let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_TWAP);
         ctx.send_transaction_with_bot(&[refresh_ix]).await.unwrap();
 
         let oracle_prices: OraclePrices = ctx.get_zero_copy_account(&feed.prices).await.unwrap();
@@ -738,7 +738,7 @@ async fn test_reset_twap() {
 
     {
         // Refresh
-        let refresh_ix = refresh_one_ix(&feed, TEST_PYTH_ORACLE);
+        let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_PYTH_ORACLE);
         ctx.send_transaction_with_bot(&[refresh_ix]).await.unwrap();
     }
 
@@ -758,7 +758,7 @@ async fn test_reset_twap() {
         token_price = Price { value: 200, exp: 6 };
         mock_oracles::set_price(&mut ctx, &feed, &TEST_PYTH_ORACLE, &token_price).await;
 
-        let refresh_ix = refresh_one_ix(&feed, TEST_PYTH_ORACLE);
+        let refresh_ix = refresh_simple_oracle_ix(&feed, TEST_PYTH_ORACLE);
         ctx.send_transaction_with_bot(&[refresh_ix]).await.unwrap();
     }
